// src/interfaces/alerts.ts
export enum AlertPositions {
  TopLeft = "TopLeft",
  TopCenter = "TopCenter",
  TopRight = "TopRight",
  BottomLeft = "BottomLeft",
  BottomCenter = "BottomCenter",
  BottomRight = "BottomRight",
  CenterLeft = "CenterLeft",
  CenterCenter = "CenterCenter",
  CenterRight = "CenterRight",
}

export interface Sound {
  soundPath?: string;
  soundDelay?: string;
  soundOffset?: string;
}

export interface Image {
  imagePath?: string;
  width?: number;
  height?: number;
  animation?: Animation;
}

export enum TextPositions {
  top = "top",
  bottom = "bottom",
  right = "right",
  left = "left",
  over = "over",
}
// ...
export enum TextStyles {
  bold = "bold",
  italic = "italic",
  underline = "underline",
}

interface Animation {
  enter?: { type?: EnterAnimationTypes; duration?: number; delay?: number };
  exit?: { type?: ExitAnimationTypes; duration?: number; offset?: number };
}

export interface Text {
  position?: TextPositions;
  content?: string;
  width?: number;
  height?: number;
  size?: string;
  color?: string;
  lineHeight?: string;
  letterSpacing?: string;
  wordSpacing?: string;
  textAlign?: string;
  textStyle?: TextStyles[];
  animation?: Animation;
  stroke?: {
    color?: string;
    width?: number;
  };
}

export interface AlertVariation {
  _id: string;
  name: string;
  backgroundColor: string;
  duration?: number;
  chances?: number;
  requiredAmount?: number;
  width?: number;
  heigth?: number;
  position?: AlertPositions;
  sound?: Sound;
  image?: Image;
  text?: Text;
  filepath?: string;
}